Change the way we are rendering message row when there should be a summary, but the summary component does not return one.
Before this change we called the summary component and check if it
returns anything, when the summary component did not returned anything
we rendered the message row instead.
Now we are just providing the message row to the message summary
component and the summary component can render it, when necessary.